requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
altair==5.3.0
annotated-types==0.7.0
attrs==23.2.0
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dataclasses-json==0.6.7
filelock==3.15.4
frozenlist==1.4.1
fsspec==2024.6.0
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.0.3
huggingface-hub==0.23.4
idna==3.7
intel-openmp==2021.4.0
Jinja2==3.1.4
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
langchain==0.2.5
langchain-community==0.2.5
langchain-core==0.2.9
langchain-text-splitters==0.2.1
langsmith==0.1.81
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.3
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
numpy==1.26.4
orjson==3.10.5
packaging==24.1
pandas==2.2.2
pillow==10.3.0
protobuf==5.27.1
pyarrow==16.1.0
pydantic==2.7.4
pydantic_core==2.18.4
pydeck==0.9.1
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
referencing==0.35.1
regex==2024.5.15
requests==2.32.3
rich==13.7.1
rpds-py==0.18.1
safetensors==0.4.3
six==1.16.0
smmap==5.0.1
SQLAlchemy==2.0.31
streamlit==1.36.0
sympy==1.12.1
tbb==2021.13.0
tenacity==8.4.1
tokenizers==0.19.1
toml==0.10.2
toolz==0.12.1
torch==2.3.1
tornado==6.4.1
tqdm==4.66.4
transformers==4.41.2
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
yarl==1.9.4
